pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "bugbounty-agent"
version = "0.1.0"
description = "AI-powered bug bounty automation framework"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "openai>=1.40.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "aiofiles>=23.0.0",
    "aiosqlite>=0.20.0",
    "jinja2>=3.1.0",
    "pyyaml>=6.0.0",
    "tldextract>=5.0.0",
    "dnspython>=2.4.0",
    "httpx>=0.27.0",
]

[project.scripts]
bugbounty = "bugbounty.main:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["bugbounty*"]
